+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 書き方は[Keep a Changelog](http://keepachangelog.com/en/1.0.0/)に基づきます。<br>
 バージョンは[Semantic Versioning](http://semver.org/spec/v2.0.0.html)に従います。
 
-## [Unreleased]
+## [0.32.0](https://github.com/worgue/magic-pocket/releases/tag/0.32.0) - 2026-08-27
 
 ### Changed
 - **破壊的**: `django-storages` を必須依存から `[django]` extra へ移しました。